New Commit - Favorite Movies List & Hist. List
This web app was created by using Promises and async functions. The new features are that for each movie you see, it get's added in a historic list just down below. Also, if you see a movie and click the heart button, it is put into your favorite movies list. Enjoy it.